+++ tensorlib/datasets/__init__.py
@@ -1,0 +1,7 @@
+"""This module deals with loading datasets."""
+from .datasets import load_bread
+from .datasets import fetch_decmeg
+
+__all__ = ['load_bread',
+           'fetch_decmeg',
+           ]
