+++ tensorlib/decomposition/__init__.py
@@ -1,0 +1,5 @@
+"""This module deals with tensor decompositions."""
+from .decomposition import cp
+
+__all__ = ['cp',
+           ]
